Вынесет в отдельный презентер логику отрисовки точки маршрута
